scratch most code (simplified data generator)
