+++ .gitignore
@@ -3,101 +3,9 @@ env
 *.egg-info
 .ipynb_checkpoints
 */results/
-*/data/*/
+*/data/*.txt
 wandb
 *.pdf
 *.zip
 *.pt
-*/data_bkp*
-*/data_cedar*
-*/plots/best_models
-*/plots/cedar_env
-*/plots/best_ever.yaml
 docs/build
-
-
-deep_continuation/small_data/B1_test.json
-deep_continuation/small_data/B1_train.json
-deep_continuation/small_data/B1_valid.json
-deep_continuation/small_data/BL_test.json
-deep_continuation/small_data/BL_train.json
-deep_continuation/small_data/BL_valid.json
-deep_continuation/small_data/Fournier_train.json
-deep_continuation/small_data/Fournier_valid.json
-deep_continuation/small_data/G1_train.json
-deep_continuation/small_data/G1_valid.json
-deep_continuation/small_data/B1/train/Pi_beta_10.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_15.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_20.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_25.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_30.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_35.0.csv
-deep_continuation/small_data/B1/train/Pi_beta_50.0.csv
-deep_continuation/small_data/B1/train/Pi.csv
-deep_continuation/small_data/B1/train/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/B1/train/SigmaRe.csv
-deep_continuation/small_data/B1/train/wmaxs.csv
-deep_continuation/small_data/B1/valid/Pi_beta_10.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_15.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_20.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_25.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_30.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_35.0.csv
-deep_continuation/small_data/B1/valid/Pi_beta_50.0.csv
-deep_continuation/small_data/B1/valid/Pi.csv
-deep_continuation/small_data/B1/valid/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/B1/valid/SigmaRe.csv
-deep_continuation/small_data/B1/valid/wmaxs.csv
-deep_continuation/small_data/BL/train/Pi_beta_30.0.csv
-deep_continuation/small_data/BL/train/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/BL/train/SigmaRe.csv
-deep_continuation/small_data/BL/train/wmaxs.csv
-deep_continuation/small_data/BL/valid/Pi_beta_30.0.csv
-deep_continuation/small_data/BL/valid/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/BL/valid/SigmaRe.csv
-deep_continuation/small_data/BL/valid/wmaxs.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_10.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_15.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_20.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_25.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_30.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_35.0.csv
-deep_continuation/small_data/Fournier/train/Pi_beta_50.0.csv
-deep_continuation/small_data/Fournier/train/Pi.csv
-deep_continuation/small_data/Fournier/train/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/Fournier/train/SigmaRe.csv
-deep_continuation/small_data/Fournier/train/wmaxs.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_10.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_15.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_20.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_25.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_30.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_35.0.csv
-deep_continuation/small_data/Fournier/valid/Pi_beta_50.0.csv
-deep_continuation/small_data/Fournier/valid/Pi.csv
-deep_continuation/small_data/Fournier/valid/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/Fournier/valid/SigmaRe.csv
-deep_continuation/small_data/Fournier/valid/wmaxs.csv
-deep_continuation/small_data/G1/train/Pi_beta_10.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_15.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_20.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_25.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_30.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_35.0.csv
-deep_continuation/small_data/G1/train/Pi_beta_50.0.csv
-deep_continuation/small_data/G1/train/Pi.csv
-deep_continuation/small_data/G1/train/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/G1/train/SigmaRe.csv
-deep_continuation/small_data/G1/train/wmaxs.csv
-deep_continuation/small_data/G1/valid/Pi_beta_10.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_15.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_20.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_25.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_30.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_35.0.csv
-deep_continuation/small_data/G1/valid/Pi_beta_50.0.csv
-deep_continuation/small_data/G1/valid/Pi.csv
-deep_continuation/small_data/G1/valid/SigmaRe_scaled_4.0.csv
-deep_continuation/small_data/G1/valid/SigmaRe.csv
-deep_continuation/small_data/G1/valid/wmaxs.csv
-
