--- deep_continuation/data/B1_valid.json
@@ -1,14 +1,0 @@
-{
-    "seed": 555,
-    "generate": 10000,
-    "path": "data/B1/valid",
-    "variant": "Beta",
-    "wmax": 20.0,
-    "beta": [10.0],
-    "nmbrs": [[0, 4],[0, 6]],
-    "cntrs": [[0.00, 0.00], [4.00, 16.0]],
-    "wdths": [[0.20, 4.00], [0.40, 4.00]],
-    "wghts": [[0.00, 1.00], [0.00, 1.00]],
-    "arngs": [[2.00, 10.0], [0.50, 10.0]],
-    "brngs": [[2.00, 10.0], [0.50, 10.0]]
-}
