--- deep_continuation/data/Fournier_train.json
@@ -1,13 +1,0 @@
-{
-    "seed": 56401,
-    "generate": 100000,
-    "path": "data/Fournier/train",
-    "variant": "Gaussian",
-    "anormal": true,
-    "wmax": 15.0,
-    "beta": [2.0],
-    "nmbrs": [[1, 1],[1, 21]],
-    "cntrs": [[0.00, 0.50], [0.00, 6.00]],
-    "wdths": [[0.10, 1.00], [0.10, 1.00]],
-    "wghts": [[1.00, 1.00], [1.00, 1.00]]
-}
