--- deep_continuation/data/Fournier_valid.json
@@ -1,13 +1,0 @@
-{
-    "seed": 124,
-    "generate": 10000,
-    "path": "data/Fournier/valid",
-    "variant": "Gaussian",
-    "anormal": true,
-    "wmax": 15.0,
-    "beta": [2.0],
-    "nmbrs": [[1, 1],[1, 21]],
-    "cntrs": [[0.00, 0.50], [0.00, 6.00]],
-    "wdths": [[0.10, 1.00], [0.10, 1.00]],
-    "wghts": [[1.00, 1.00], [1.00, 1.00]]
-}
