--- deep_continuation/data/G1_valid.json
@@ -1,12 +1,0 @@
-{
-    "seed": 111,
-    "generate": 10000,
-    "path": "data/G1/valid",
-    "variant": "Gaussian",
-    "wmax": 20.0,
-    "beta": [10.0],
-    "nmbrs": [[0, 4],[0, 6]],
-    "cntrs": [[0.00, 0.00], [4.00, 16.0]],
-    "wdths": [[0.20, 4.00], [0.40, 4.00]],
-    "wghts": [[0.00, 1.00], [0.00, 1.00]]
-}
