--- deep_continuation/plots/Pi_figure_corrected.csv
@@ -1,1 +1,0 @@
-1.000000000000000000e+00,9.203332662582397461e-01,8.856627941131591797e-01,8.618347048759460449e-01,8.411887288093566895e-01,8.215928673744201660e-01,8.023528456687927246e-01,7.832318544387817383e-01,7.641634941101074219e-01,7.451510429382324219e-01,7.262274622917175293e-01,7.074368596076965332e-01,6.888261437416076660e-01,6.704401373863220215e-01,6.523196101188659668e-01,6.345002651214599609e-01,6.170124411582946777e-01,5.998810529708862305e-01,5.831261873245239258e-01,5.667632222175598145e-01,5.508035421371459961e-01,5.352548956871032715e-01,5.201218128204345703e-01,5.054063200950622559e-01,4.911078810691833496e-01,4.772242605686187744e-01,4.637514352798461914e-01,4.506840705871582031e-01,4.380157589912414551e-01,4.257391095161437988e-01,4.138461053371429443e-01,4.023281335830688477e-01,3.911761045455932617e-01,3.803807199001312256e-01,3.699323236942291260e-01,3.598213195800781250e-01,3.500378131866455078e-01,3.405721187591552734e-01,3.314144611358642578e-01,3.225552141666412354e-01,3.139848709106445312e-01,3.056939840316772461e-01,2.976733744144439697e-01,2.899140417575836182e-01,2.824071943759918213e-01,2.751441001892089844e-01,2.681165337562561035e-01,2.613162100315093994e-01,2.547353208065032959e-01,2.483661174774169922e-01,2.422011494636535645e-01,2.362332046031951904e-01,2.304553091526031494e-01,2.248606979846954346e-01,2.194428741931915283e-01,2.141954898834228516e-01,2.091124951839447021e-01,2.041880339384078979e-01,1.994164139032363892e-01,1.947921812534332275e-01,1.903101205825805664e-01,1.859651207923889160e-01,1.817523241043090820e-01,1.776670366525650024e-01,1.737047135829925537e-01,1.698610335588455200e-01,1.661317795515060425e-01,1.625129431486129761e-01,1.590006351470947266e-01,1.555911302566528320e-01,1.522808372974395752e-01,1.490662842988967896e-01,1.459441930055618286e-01,1.429113447666168213e-01,1.399646550416946411e-01,1.371011883020401001e-01,1.343180984258651733e-01,1.316126286983489990e-01,1.289821863174438477e-01,1.264242082834243774e-01,1.239362806081771851e-01,1.215160414576530457e-01,1.191612631082534790e-01,1.168697476387023926e-01,1.146394312381744385e-01,1.124682798981666565e-01,1.103543937206268311e-01,1.082958802580833435e-01,1.062909588217735291e-01,1.043379083275794983e-01,1.024350598454475403e-01,1.005808189511299133e-01,9.877364337444305420e-02,9.701204299926757812e-02,9.529458731412887573e-02,9.361990541219711304e-02,9.198665618896484375e-02,9.039357304573059082e-02,8.883940428495407104e-02,8.732297271490097046e-02,8.584313094615936279e-02,8.439876139163970947e-02,8.298880606889724731e-02,8.161223679780960083e-02,8.026804029941558838e-02,7.895527780055999756e-02,7.767301797866821289e-02,7.642035931348800659e-02,7.519645243883132935e-02,7.400044798851013184e-02,7.283154875040054321e-02,7.168898731470108032e-02,7.057200372219085693e-02,6.947986781597137451e-02,6.841187924146652222e-02,6.736736744642257690e-02,6.634566932916641235e-02,6.534615904092788696e-02,6.436821818351745605e-02,6.341125816106796265e-02,6.247470527887344360e-02,6.155800446867942810e-02,6.066061556339263916e-02,5.978202447295188904e-02,5.892172828316688538e-02,5.807923898100852966e-02,5.725408345460891724e-02,5.644580349326133728e-02
