--- deep_continuation/plots/Pol.txt
@@ -1,128 +1,0 @@
-1.047197551196597659e-01 1.000632467815854021e+00
-3.141592653589793116e-01 9.627397803897349560e-01
-5.235987755982988157e-01 9.274481274155266242e-01
-7.330382858376184307e-01 8.943338128769013906e-01
-9.424777960769379348e-01 8.630605543792384537e-01
-1.151917306316257328e+00 8.333638290167584284e-01
-1.361356816555577165e+00 8.050372518367908681e-01
-1.570796326794896336e+00 7.779208634968199920e-01
-1.780235837034216173e+00 7.518912283040102285e-01
-1.989675347273535566e+00 7.268532210122601311e-01
-2.199114857512855181e+00 7.027333434731688477e-01
-2.408554367752174574e+00 6.794743827750057141e-01
-2.617993877991494411e+00 6.570312102679761734e-01
-2.827433388230813804e+00 6.353675257020964251e-01
-3.036872898470133642e+00 6.144533677059712540e-01
-3.246312408709452590e+00 5.942632352069385782e-01
-3.455751918948772428e+00 5.747746894421452701e-01
-3.665191429188091821e+00 5.559673299383274259e-01
-3.874630939427411658e+00 5.378220587423462318e-01
-4.084070449666731051e+00 5.203205647922384269e-01
-4.293509959906050000e+00 5.034449747443562329e-01
-4.502949470145369837e+00 4.871776281986909263e-01
-4.712388980384689674e+00 4.715009445497427842e-01
-4.921828490624009511e+00 4.563973560705562038e-01
-5.131268000863328460e+00 4.418492876883218234e-01
-5.340707511102648297e+00 4.278391685388455712e-01
-5.550147021341968134e+00 4.143494640402110574e-01
-5.759586531581287971e+00 4.013627200984378196e-01
-5.969026041820606920e+00 3.888616133051832624e-01
-6.178465552059926758e+00 3.768290027337770076e-01
-6.387905062299245706e+00 3.652479802852396928e-01
-6.597344572538565544e+00 3.541019175614762982e-01
-6.806784082777884493e+00 3.433745080145160777e-01
-7.016223593017204330e+00 3.330498036920613858e-01
-7.225663103256524167e+00 3.231122463142778178e-01
-7.435102613495844004e+00 3.135466927098751744e-01
-7.644542123735162953e+00 3.043384348394062022e-01
-7.853981633974482790e+00 2.954732147629416383e-01
-8.063421144213801739e+00 2.869372349857750581e-01
-8.272860654453122464e+00 2.787171646536252823e-01
-8.482300164692441413e+00 2.708001420788808677e-01
-8.691739674931762138e+00 2.631737740701782235e-01
-8.901179185171081087e+00 2.558261325154352339e-01
-9.110618695410400036e+00 2.487457486381751404e-01
-9.320058205649720762e+00 2.419216053121109877e-01
-9.529497715889039711e+00 2.353431277821176615e-01
-9.738937226128358660e+00 2.290001731026784504e-01
-9.948376736367679385e+00 2.228830185689475718e-01
-1.015781624660699656e+01 2.169823493814941606e-01
-1.036725575684631728e+01 2.112892457541119817e-01
-1.057669526708563623e+01 2.057951696450416312e-01
-1.078613477732495518e+01 2.004919512656480052e-01
-1.099557428756427591e+01 1.953717754970140386e-01
-1.120501379780359485e+01 1.904271683239462387e-01
-1.141445330804291380e+01 1.856509833773788698e-01
-1.162389281828223453e+01 1.810363886599450911e-01
-1.183333232852155348e+01 1.765768535153494634e-01
-1.204277183876087420e+01 1.722661358899361117e-01
-1.225221134900019315e+01 1.680982699242964951e-01
-1.246165085923951210e+01 1.640675539037297348e-01
-1.267109036947883283e+01 1.601685385886663093e-01
-1.288052987971815178e+01 1.563960159396470817e-01
-1.308996938995747072e+01 1.527450082459551883e-01
-1.329940890019679145e+01 1.492107576624063792e-01
-1.350884841043611040e+01 1.457887161549893951e-01
-1.371828792067543112e+01 1.424745358529071193e-01
-1.392772743091475007e+01 1.392640598020111753e-01
-1.413716694115406902e+01 1.361533131125586271e-01
-1.434660645139338975e+01 1.331384944925867175e-01
-1.455604596163270870e+01 1.302159681569259320e-01
-1.476548547187202765e+01 1.273822561009055332e-01
-1.497492498211134837e+01 1.246340307271011499e-01
-1.518436449235066554e+01 1.219681078129833535e-01
-1.539380400258998627e+01 1.193814398070194988e-01
-1.560324351282930522e+01 1.168711094406288259e-01
-1.581268302306862417e+01 1.144343236433580935e-01
-1.602212253330794667e+01 1.120684077487178881e-01
-1.623156204354726384e+01 1.097707999782764066e-01
-1.644100155378658457e+01 1.075390461918283441e-01
-1.665044106402590529e+01 1.053707948917299991e-01
-1.685988057426522246e+01 1.032637924698084630e-01
-1.706932008450454319e+01 1.012158786855965192e-01
-1.727875959474386036e+01 9.922498236501446633e-02
-1.748819910498318464e+01 9.728911730900127730e-02
-1.769763861522250181e+01 9.540637840199200914e-02
-1.790707812546181898e+01 9.357493791053361765e-02
-1.811651763570113971e+01 9.179304196273051220e-02
-1.832595714594045688e+01 9.005900719960555079e-02
-1.853539665617978116e+01 8.837121758985169717e-02
-1.874483616641909833e+01 8.672812139983308855e-02
-1.895427567665841906e+01 8.512822831106643673e-02
-1.916371518689773623e+01 8.357010667777800550e-02
-1.937315469713705696e+01 8.205238091748326412e-02
-1.958259420737637768e+01 8.057372902787585112e-02
-1.979203371761569841e+01 7.913288022364170504e-02
-2.000147322785501558e+01 7.772861268712794036e-02
-2.021091273809433631e+01 7.635975142710045571e-02
-2.042035224833365348e+01 7.502516624011223989e-02
-2.062979175857297420e+01 7.372376976928385339e-02
-2.083923126881229493e+01 7.245451565555996454e-02
-2.104867077905161565e+01 7.121639677676123992e-02
-2.125811028929093283e+01 7.000844356998905949e-02
-2.146754979953025355e+01 6.882972243317261873e-02
-2.167698930976957072e+01 6.767933420176359960e-02
-2.188642882000889500e+01 6.655641269679299366e-02
-2.209586833024821217e+01 6.546012334070103189e-02
-2.230530784048753290e+01 6.438966183753720707e-02
-2.251474735072685007e+01 6.334425291430773042e-02
-2.272418686096616725e+01 6.232314912041251459e-02
-2.293362637120549152e+01 6.132562968227708855e-02
-2.314306588144480870e+01 6.035099941043347432e-02
-2.335250539168412942e+01 5.939858765644887711e-02
-2.356194490192344659e+01 5.846774731723687613e-02
-2.377138441216276732e+01 5.755785388441376355e-02
-2.398082392240208449e+01 5.666830453648544902e-02
-2.419026343264140877e+01 5.579851727176498438e-02
-2.439970294288072594e+01 5.494793008003047430e-02
-2.460914245312004667e+01 5.411600015103597283e-02
-2.481858196335936384e+01 5.330220311808675504e-02
-2.502802147359868457e+01 5.250603233498151584e-02
-2.523746098383800529e+01 5.172699818471336586e-02
-2.544690049407732602e+01 5.096462741840272759e-02
-2.565634000431664319e+01 5.021846252301528230e-02
-2.586577951455596391e+01 4.948806111649097644e-02
-2.607521902479528109e+01 4.877299536898141441e-02
-2.628465853503460536e+01 4.807285144895897394e-02
-2.649409804527392254e+01 4.738722899302409697e-02
-2.670353755551324326e+01 4.671574059829645270e-02
