--- deep_continuation/plots/figure_optimal_spectral_function_tem0.05_alpha1.59e+07.dat
@@ -1,114 +1,0 @@
-   0.000000000000e+00   1.633267421393e-01
-   1.205925415597e-01   1.642682502379e-01
-   2.411850831194e-01   1.656211986215e-01
-   3.617776246792e-01   1.658128315784e-01
-   4.823701662389e-01   1.649418897451e-01
-   6.029627077986e-01   1.635660756732e-01
-   7.235552493583e-01   1.621115386861e-01
-   8.441477909180e-01   1.608253789057e-01
-   9.647403324778e-01   1.598308576536e-01
-   1.085332874037e+00   1.591767960889e-01
-   1.205925415597e+00   1.588691580697e-01
-   1.326517957157e+00   1.588900154467e-01
-   1.447110498717e+00   1.592087393476e-01
-   1.567703040276e+00   1.597885786014e-01
-   1.688295581836e+00   1.605904989783e-01
-   1.808888123396e+00   1.615753958782e-01
-   1.929480664956e+00   1.627053417866e-01
-   2.050073206515e+00   1.639442642485e-01
-   2.170665748075e+00   1.652582911489e-01
-   2.291258289635e+00   1.666159046455e-01
-   2.411850831194e+00   1.679879873068e-01
-   2.532443372754e+00   1.693478092939e-01
-   2.653035914314e+00   1.706709843763e-01
-   2.773628455874e+00   1.719354102673e-01
-   2.894220997433e+00   1.731212011324e-01
-   3.014813538993e+00   1.742106163599e-01
-   3.135406080553e+00   1.751879869166e-01
-   3.255998622112e+00   1.760396395465e-01
-   3.376591163672e+00   1.767538192570e-01
-   3.497183705232e+00   1.773206084305e-01
-   3.617776246792e+00   1.777318442205e-01
-   3.738368788351e+00   1.779810319537e-01
-   3.858961329911e+00   1.780632565535e-01
-   3.979553871471e+00   1.779750903430e-01
-   4.100146413031e+00   1.777145003861e-01
-   4.220738954590e+00   1.772807507360e-01
-   4.341331496150e+00   1.766743086462e-01
-   4.461924037710e+00   1.758967442401e-01
-   4.582516579269e+00   1.749506360011e-01
-   4.703109120829e+00   1.738394729936e-01
-   4.823701662389e+00   1.725675612233e-01
-   4.944294203949e+00   1.711399305460e-01
-   5.064886745508e+00   1.695622452139e-01
-   5.185479287068e+00   1.678407172132e-01
-   5.306071828628e+00   1.659820235869e-01
-   5.426664370187e+00   1.639932265905e-01
-   5.547256911747e+00   1.618817009376e-01
-   5.667849453307e+00   1.596550605779e-01
-   5.788441994867e+00   1.573210970397e-01
-   5.909034536426e+00   1.548877148303e-01
-   6.029627077986e+00   1.523628784146e-01
-   6.150219619546e+00   1.497545589028e-01
-   6.270812161106e+00   1.470706889640e-01
-   6.391404702665e+00   1.443191194708e-01
-   6.511997244225e+00   1.415075837702e-01
-   6.632589785785e+00   1.386436611648e-01
-   6.753182327344e+00   1.357347533347e-01
-   6.873774868904e+00   1.327880513245e-01
-   6.994367410464e+00   1.298105224017e-01
-   7.114959952024e+00   1.268088843889e-01
-   7.235552493583e+00   1.237895959507e-01
-   7.356145035143e+00   1.207588426079e-01
-   7.476737576703e+00   1.177225259462e-01
-   7.597330118262e+00   1.146862615246e-01
-   7.717922659822e+00   1.116553682038e-01
-   7.838515201382e+00   1.086348722621e-01
-   7.959107742942e+00   1.056295013005e-01
-   8.079700284501e+00   1.026436918570e-01
-   8.200292826061e+00   9.968157679194e-02
-   8.320885367621e+00   9.674704079334e-02
-   8.441477909180e+00   9.384354491994e-02
-   8.562070450740e+00   9.097478086671e-02
-   8.682662992300e+00   8.814233014165e-02
-   8.803255533860e+00   8.535417812787e-02
-   8.923848075419e+00   8.262220650977e-02
-   9.044440616979e+00   7.980099955303e-02
-   9.165033158539e+00   7.719645397845e-02
-   9.292144215859e+00   7.448945309392e-02
-   9.426316998585e+00   7.164682371804e-02
-   9.568156797467e+00   6.871644757535e-02
-   9.718340113930e+00   6.569301878518e-02
-   9.877625449574e+00   6.257896987591e-02
-   1.004686611869e+01   5.937611342411e-02
-   1.022702554066e+01   5.608794222846e-02
-   1.041919559076e+01   5.271938855812e-02
-   1.062461874776e+01   4.927724297954e-02
-   1.084471498741e+01   4.577046383638e-02
-   1.108111465221e+01   4.221056189031e-02
-   1.133569890661e+01   3.861201011496e-02
-   1.161064990137e+01   3.499268011522e-02
-   1.190851347902e+01   3.137428157438e-02
-   1.223227823734e+01   2.778277350295e-02
-   1.258547615551e+01   2.424869396010e-02
-   1.297231197064e+01   2.080733065074e-02
-   1.339783136729e+01   1.749861598821e-02
-   1.386814227937e+01   1.436659512183e-02
-   1.439070995946e+01   1.145824092266e-02
-   1.497475619015e+01   8.821413786298e-03
-   1.563180819968e+01   6.501564776432e-03
-   1.637646714381e+01   4.537312663713e-03
-   1.722750593710e+01   2.954087787816e-03
-   1.820947377552e+01   1.758218888964e-03
-   1.935510292034e+01   9.288812132211e-04
-   2.070902827330e+01   4.182644128675e-04
-   2.233373869686e+01   1.499259508417e-04
-   2.431949588121e+01   3.930922549478e-05
-   2.680169236165e+01   6.125074500676e-06
-   2.999308783650e+01   4.855729102409e-07
-   3.424828180296e+01   7.988935275680e-09
-   4.020555335601e+01   2.393880956032e-11
-   4.914146068559e+01   8.438004184099e-17
-   6.403463956821e+01   7.222798749127e-21
-   9.382099733346e+01   2.980908959686e-23
-   1.831800706292e+02   0.000000000000e+00
