--- deep_continuation/plots/optimal_spectral_function_tem0.03333_alpha1.59e+07.dat
@@ -1,114 +1,0 @@
-   0.000000000000e+00   2.449899529967e-01
-   8.039503061507e-02   2.464022334039e-01
-   1.607900612301e-01   2.484316891193e-01
-   2.411850918452e-01   2.487191624154e-01
-   3.215801224603e-01   2.474127635478e-01
-   4.019751530753e-01   2.453490516079e-01
-   4.823701836904e-01   2.431672538206e-01
-   5.627652143055e-01   2.412380215704e-01
-   6.431602449205e-01   2.397462470626e-01
-   7.235552755356e-01   2.387651619304e-01
-   8.039503061507e-01   2.383037118003e-01
-   8.843453367657e-01   2.383350043178e-01
-   9.647403673808e-01   2.388130960883e-01
-   1.045135397996e+00   2.396828603187e-01
-   1.125530428611e+00   2.408857456440e-01
-   1.205925459226e+00   2.423630951745e-01
-   1.286320489841e+00   2.440580176571e-01
-   1.366715520456e+00   2.459164044320e-01
-   1.447110551071e+00   2.478874473658e-01
-   1.527505581686e+00   2.499238697162e-01
-   1.607900612301e+00   2.519819953784e-01
-   1.688295642916e+00   2.540217296301e-01
-   1.768690673531e+00   2.560064931448e-01
-   1.849085704146e+00   2.579031325331e-01
-   1.929480734762e+00   2.596818190980e-01
-   2.009875765377e+00   2.613159418927e-01
-   2.090270795992e+00   2.627819974482e-01
-   2.170665826607e+00   2.640594759402e-01
-   2.251060857222e+00   2.651307447460e-01
-   2.331455887837e+00   2.659809277022e-01
-   2.411850918452e+00   2.665977803426e-01
-   2.492245949067e+00   2.669715609277e-01
-   2.572640979682e+00   2.670948963223e-01
-   2.653036010297e+00   2.669626461124e-01
-   2.733431040912e+00   2.665717593568e-01
-   2.813826071527e+00   2.659211337878e-01
-   2.894221102142e+00   2.650114690849e-01
-   2.974616132757e+00   2.638451209720e-01
-   3.055011163372e+00   2.624259571388e-01
-   3.135406193988e+00   2.607592112321e-01
-   3.215801224603e+00   2.588513421877e-01
-   3.296196255218e+00   2.567098943908e-01
-   3.376591285833e+00   2.543433655829e-01
-   3.456986316448e+00   2.517610717094e-01
-   3.537381347063e+00   2.489730302122e-01
-   3.617776377678e+00   2.459898336422e-01
-   3.698171408293e+00   2.428225435878e-01
-   3.778566438908e+00   2.394825823343e-01
-   3.858961469523e+00   2.359816356603e-01
-   3.939356500138e+00   2.323315617325e-01
-   4.019751530753e+00   2.285443056655e-01
-   4.100146561368e+00   2.246318261300e-01
-   4.180541591983e+00   2.206060199960e-01
-   4.260936622598e+00   2.164786655250e-01
-   4.341331653214e+00   2.122613605329e-01
-   4.421726683829e+00   2.079654772513e-01
-   4.502121714444e+00   2.036021137208e-01
-   4.582516745059e+00   1.991820614270e-01
-   4.662911775674e+00   1.947157666533e-01
-   4.743306806289e+00   1.902133101116e-01
-   4.823701836904e+00   1.856843769715e-01
-   4.904096867519e+00   1.811382461980e-01
-   4.984491898134e+00   1.765837722778e-01
-   5.064886928749e+00   1.720293740519e-01
-   5.145281959364e+00   1.674830353130e-01
-   5.225676989979e+00   1.629522906366e-01
-   5.306072020594e+00   1.584442347701e-01
-   5.386467051209e+00   1.539655196957e-01
-   5.466862081824e+00   1.495223484895e-01
-   5.547257112439e+00   1.451205441770e-01
-   5.627652143055e+00   1.407652997746e-01
-   5.708047173670e+00   1.364621552174e-01
-   5.788442204285e+00   1.322134779501e-01
-   5.868837234900e+00   1.280312513071e-01
-   5.949232265515e+00   1.239332941226e-01
-   6.029627296130e+00   1.197014830566e-01
-   6.110022326745e+00   1.157946658312e-01
-   6.194763034691e+00   1.117341646066e-01
-   6.284211559744e+00   1.074702207862e-01
-   6.378771429087e+00   1.030746572249e-01
-   6.478893643685e+00   9.853951424137e-02
-   6.585083871289e+00   9.386844153575e-02
-   6.697910988118e+00   8.906415729974e-02
-   6.818017273774e+00   8.413190105600e-02
-   6.946130645142e+00   7.907907121053e-02
-   7.083079421431e+00   7.391585341933e-02
-   7.229810253169e+00   6.865568540606e-02
-   7.387410035407e+00   6.331583318763e-02
-   7.557132877816e+00   5.791800625915e-02
-   7.740433547618e+00   5.248901203099e-02
-   7.939009273238e+00   4.706141500051e-02
-   8.154852453259e+00   4.167415368944e-02
-   8.390317740554e+00   3.637303518452e-02
-   8.648208293306e+00   3.121099101411e-02
-   8.931887901334e+00   2.624791980040e-02
-   9.245428520732e+00   2.154988924964e-02
-   9.593806986731e+00   1.718735864857e-02
-   9.983171154612e+00   1.323211858427e-02
-   1.042120584348e+01   9.752345631024e-03
-   1.091764515753e+01   6.805967937240e-03
-   1.148500437358e+01   4.431131002558e-03
-   1.213964962287e+01   2.637327937177e-03
-   1.290340241372e+01   1.393321615540e-03
-   1.380601934835e+01   6.273965301752e-04
-   1.488915966991e+01   2.248888955317e-04
-   1.621299784070e+01   5.896383055223e-05
-   1.786779555420e+01   9.187610624836e-06
-   1.999539261440e+01   7.283592812660e-07
-   2.283218869468e+01   1.198340169295e-08
-   2.680370320706e+01   3.590821072643e-11
-   3.276097497564e+01   1.265695170505e-16
-   4.268976125660e+01   7.222823001914e-21
-   6.254733381852e+01   2.980909396328e-23
-   1.221200515043e+02   0.000000000000e+00
