--- deep_continuation/plots/optimal_spectral_function_tem0.05_alpha1.73e+06.dat
@@ -1,114 +1,0 @@
-   0.000000000000e+00   3.853855863928e-01
-   1.328655811054e-01   3.600286605468e-01
-   2.657311622108e-01   2.968571965188e-01
-   3.985967433162e-01   2.272926706384e-01
-   5.314623244216e-01   1.722651147453e-01
-   6.643279055270e-01   1.343035625159e-01
-   7.971934866324e-01   1.092647386139e-01
-   9.300590677378e-01   9.289107042138e-02
-   1.062924648843e+00   8.219826631035e-02
-   1.195790229949e+00   7.529213967810e-02
-   1.328655811054e+00   7.099404455604e-02
-   1.461521392159e+00   6.856357966525e-02
-   1.594386973265e+00   6.752574334301e-02
-   1.727252554370e+00   6.756780006898e-02
-   1.860118135476e+00   6.847782685915e-02
-   1.992983716581e+00   7.010752757424e-02
-   2.125849297686e+00   7.234923194152e-02
-   2.258714878792e+00   7.512132771164e-02
-   2.391580459897e+00   7.835882449025e-02
-   2.524446041003e+00   8.200712568145e-02
-   2.657311622108e+00   8.601786510686e-02
-   2.790177203213e+00   9.034611482120e-02
-   2.923042784319e+00   9.494853234528e-02
-   3.055908365424e+00   9.978217230026e-02
-   3.188773946530e+00   1.048037808246e-01
-   3.321639527635e+00   1.099694487202e-01
-   3.454505108740e+00   1.152345346024e-01
-   3.587370689846e+00   1.205537922183e-01
-   3.720236270951e+00   1.258816481188e-01
-   3.853101852057e+00   1.311725892743e-01
-   3.985967433162e+00   1.363816221693e-01
-   4.118833014267e+00   1.414647740321e-01
-   4.251698595373e+00   1.463796081442e-01
-   4.384564176478e+00   1.510857321680e-01
-   4.517429757583e+00   1.555452753582e-01
-   4.650295338689e+00   1.597233285703e-01
-   4.783160919794e+00   1.635883186442e-01
-   4.916026500900e+00   1.671123280375e-01
-   5.048892082005e+00   1.702713350183e-01
-   5.181757663110e+00   1.730453818544e-01
-   5.314623244216e+00   1.754186750330e-01
-   5.447488825321e+00   1.773795955840e-01
-   5.580354406427e+00   1.789206671548e-01
-   5.713219987532e+00   1.800384305185e-01
-   5.846085568637e+00   1.807332894065e-01
-   5.978951149743e+00   1.810092867790e-01
-   6.111816730848e+00   1.808738555783e-01
-   6.244682311954e+00   1.803375254984e-01
-   6.377547893059e+00   1.794136155810e-01
-   6.510413474164e+00   1.781179036164e-01
-   6.643279055270e+00   1.764682919301e-01
-   6.776144636375e+00   1.744844690308e-01
-   6.909010217481e+00   1.721875726844e-01
-   7.041875798586e+00   1.695998774286e-01
-   7.174741379691e+00   1.667444678040e-01
-   7.307606960797e+00   1.636449625973e-01
-   7.440472541902e+00   1.603252333789e-01
-   7.573338123008e+00   1.568091622238e-01
-   7.706203704113e+00   1.531204120656e-01
-   7.839069285218e+00   1.492822288201e-01
-   7.971934866324e+00   1.453172687720e-01
-   8.104800447429e+00   1.412474425880e-01
-   8.237666028535e+00   1.370937928832e-01
-   8.370531609640e+00   1.328763925467e-01
-   8.503397190745e+00   1.286142481382e-01
-   8.636262771851e+00   1.243252667050e-01
-   8.769128352956e+00   1.200261700853e-01
-   8.901993934062e+00   1.157325225942e-01
-   9.034859515167e+00   1.114586253522e-01
-   9.167725096272e+00   1.072177748745e-01
-   9.300590677378e+00   1.030213973468e-01
-   9.433456258483e+00   9.888220519869e-02
-   9.566321839589e+00   9.480299858039e-02
-   9.699187420694e+00   9.081693032217e-02
-   9.832053001799e+00   8.691201992477e-02
-   9.964918582905e+00   8.287324325151e-02
-   1.009778416401e+01   7.923815822238e-02
-   1.023783166842e+01   7.546742922176e-02
-   1.038565958974e+01   7.154476005761e-02
-   1.054193482085e+01   6.754064298336e-02
-   1.070740271261e+01   6.346407176141e-02
-   1.088289896145e+01   5.932675868926e-02
-   1.106936372584e+01   5.514256334981e-02
-   1.126785847504e+01   5.092789513647e-02
-   1.147958620751e+01   4.670185953003e-02
-   1.170591585256e+01   4.248642605004e-02
-   1.194841190084e+01   3.830652145441e-02
-   1.220887061935e+01   3.419004071966e-02
-   1.248936462391e+01   3.016773364156e-02
-   1.279229814883e+01   2.627292609401e-02
-   1.312047613416e+01   2.254102598260e-02
-   1.347719133560e+01   1.900875800555e-02
-   1.386633519173e+01   1.571307031432e-02
-   1.429254036748e+01   1.268966589989e-02
-   1.476136606081e+01   9.971115715774e-03
-   1.527954182712e+01   7.584595291728e-03
-   1.585529267858e+01   5.549227740197e-03
-   1.649877892432e+01   3.873449977998e-03
-   1.722270095079e+01   2.552154959499e-03
-   1.804314591411e+01   1.565461294922e-03
-   1.898079730077e+01   8.768920807411e-04
-   2.006270274691e+01   4.377018464133e-04
-   2.132492576742e+01   1.868763796449e-04
-   2.281664388255e+01   6.563414729258e-05
-   2.460670562072e+01   1.676203285640e-05
-   2.679455885626e+01   3.141397711651e-06
-   2.952937540067e+01   2.360749126969e-07
-   3.304556810064e+01   1.600764818280e-08
-   3.773382503393e+01   1.909072517176e-11
-   4.429738474054e+01   6.677675151273e-13
-   5.414272430045e+01   6.304028133150e-17
-   7.055162356697e+01   2.992664720565e-21
-   1.033694221000e+02  -1.606273372572e-14
-   2.018228176991e+02   0.000000000000e+00
