--- docs/source/deep_continuation.rst
@@ -1,85 +1,0 @@
-deep\_continuation package
-==========================
-
-Submodules
-----------
-
-deep\_continuation.data module
-------------------------------
-
-.. automodule:: deep_continuation.data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.data\_generator module
------------------------------------------
-
-.. automodule:: deep_continuation.data_generator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.function\_generator module
----------------------------------------------
-
-.. automodule:: deep_continuation.function_generator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.infinite\_data module
-----------------------------------------
-
-.. automodule:: deep_continuation.infinite_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.monotonous\_functions module
------------------------------------------------
-
-.. automodule:: deep_continuation.monotonous_functions
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.random\_search module
-----------------------------------------
-
-.. automodule:: deep_continuation.random_search
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.train module
--------------------------------
-
-.. automodule:: deep_continuation.train
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.utils module
--------------------------------
-
-.. automodule:: deep_continuation.utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-deep\_continuation.wandb\_utils module
---------------------------------------
-
-.. automodule:: deep_continuation.wandb_utils
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: deep_continuation
-   :members:
-   :undoc-members:
-   :show-inheritance:
