--- docs/source/index.rst
@@ -1,20 +1,0 @@
-.. Deep Continuation documentation master file, created by
-   sphinx-quickstart on Sat Feb 12 23:43:58 2022.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to Deep Continuation's documentation!
-=============================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
