--- docs/source/modules.rst
@@ -1,8 +1,0 @@
-deep_continuation
-=================
-
-.. toctree::
-   :maxdepth: 4
-
-   deep_continuation
-   setup
